+++ CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0-rc.0](https://github.com/umbraco/Umbraco.UI/compare/v1.5.0...v1.6.0-rc.0) (2023-11-08)
+
+### Bug Fixes
+
+- correct green colors ([#648](https://github.com/umbraco/Umbraco.UI/issues/648)) ([d33d4e6](https://github.com/umbraco/Umbraco.UI/commit/d33d4e60b04dfd6873abd7062629de4559a1d3d5))
+
+### Features
+
+- Don't hide placeholder text on focus ([#624](https://github.com/umbraco/Umbraco.UI/issues/624)) ([9cb90b4](https://github.com/umbraco/Umbraco.UI/commit/9cb90b40504e47997756eebb3955bb016057f6e0))
+- **uui-popover-container:** firefox polyfill ([#628](https://github.com/umbraco/Umbraco.UI/issues/628)) ([7242825](https://github.com/umbraco/Umbraco.UI/commit/72428255cc67ceeb4af8564efe051d99098b795f))
+- **uui-swatch:** ability to overwrite displayed color with a css custom prop ([#649](https://github.com/umbraco/Umbraco.UI/issues/649)) ([48db9f5](https://github.com/umbraco/Umbraco.UI/commit/48db9f5d2b6e1bb59be430506b757e99ed8a0ea2))
+
+# [1.5.0](https://github.com/umbraco/Umbraco.UI/compare/v1.5.0-rc.3...v1.5.0) (2023-10-30)
+
+**Note:** Version bump only for package uui-monorepo
+
+# [1.5.0-rc.3](https://github.com/umbraco/Umbraco.UI/compare/v1.5.0-rc.2...v1.5.0-rc.3) (2023-10-18)
+
+**Note:** Version bump only for package uui-monorepo
+
 # [1.5.0-rc.2](https://github.com/umbraco/Umbraco.UI/compare/v1.5.0-rc.1...v1.5.0-rc.2) (2023-10-18)
 
 **Note:** Version bump only for package uui-monorepo