+++ packages/uui-css/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0-rc.0](https://github.com/umbraco/Umbraco.UI/compare/v1.5.0...v1.6.0-rc.0) (2023-11-08)
+
+### Bug Fixes
+
+- correct green colors ([#648](https://github.com/umbraco/Umbraco.UI/issues/648)) ([d33d4e6](https://github.com/umbraco/Umbraco.UI/commit/d33d4e60b04dfd6873abd7062629de4559a1d3d5))
+
+# [1.5.0](https://github.com/umbraco/Umbraco.UI/compare/v1.5.0-rc.3...v1.5.0) (2023-10-30)
+
+**Note:** Version bump only for package @umbraco-ui/uui-css
+
 # [1.4.0](https://github.com/umbraco/Umbraco.UI/compare/v1.4.0-rc.2...v1.4.0) (2023-09-21)
 
 **Note:** Version bump only for package @umbraco-ui/uui-css