+++ packages/uui-toggle/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0-rc.0](https://github.com/umbraco/Umbraco.UI/compare/v1.5.0...v1.6.0-rc.0) (2023-11-08)
+
+**Note:** Version bump only for package @umbraco-ui/uui-toggle
+
+# [1.5.0](https://github.com/umbraco/Umbraco.UI/compare/v1.5.0-rc.3...v1.5.0) (2023-10-30)
+
+**Note:** Version bump only for package @umbraco-ui/uui-toggle
+
+# [1.5.0-rc.3](https://github.com/umbraco/Umbraco.UI/compare/v1.5.0-rc.2...v1.5.0-rc.3) (2023-10-18)
+
+**Note:** Version bump only for package @umbraco-ui/uui-toggle
+
 # [1.5.0-rc.2](https://github.com/umbraco/Umbraco.UI/compare/v1.5.0-rc.1...v1.5.0-rc.2) (2023-10-18)
 
 **Note:** Version bump only for package @umbraco-ui/uui-toggle